Update TOP EARNERS.sql
I came that much closer to solving this problem known as "top earners" on hacker rank dot com.

I re-learned a detail about aliases ("AS") and that they only last the duration of the query in which they are in.  This solved the previous issue plus paved a new way.