09 08 2024
Managed to add the new column but I am having problems populating it.

Somehow a column that I am not accounting for is important.  How is employee_id needed in an operation between the column 'months' and 'salary'?

I will try to delete the newly added column 'monthsxsalary' next time and added it as NOT NULL next time.